+++ apps/client/src/features/watching/index.ts
@@ -1,3 +1,2 @@
-export { useConsumer } from './model';
-
 export { LiveCamperInfo, LivePlayer } from './ui';
+export { useConsume } from './model';
