+++ apps/client/src/features/watching/model/index.ts
@@ -1,1 +1,1 @@
-export { useConsumer } from './useConsumer';
+export { useConsume } from './useConsume';
