+++ apps/client/src/shared/lib/index.ts
@@ -1,5 +1,5 @@
 export { useSocket } from './useSocket';
 export { useTheme } from './useTheme';
 export { useToast } from './useToast';
-export { useTransport } from './useTransport';
+export { getRtpCapabilities, createDevice, connectTransport } from './mediasoupHelpers';
 export { cn, checkDependencies } from './utils';
